add position and length utilization; also a bunch of unit tests around it

+++ repo/android-av/src/com/steelthorn/android/av/IScanDefinition.java
@@ -1,6 +1,5 @@
 package com.steelthorn.android.av;
 
-import java.util.List;
 
 public interface IScanDefinition
 {
@@ -8,7 +7,7 @@ public interface IScanDefinition
 	
 	byte[] getHashValue();
 	
-	long getMatchSize();
+	int getMatchSize();
 	
 	long getPosition();
